refactor url input processing logic to parent component
