+++ twitch-spy-frontend/src/components/URLInput.tsx
@@ -26,6 +26,29 @@ export const URLInput = () => {
     }
   };
 
+  const cleanInput = (input: string): string => {
+    return input
+      .split(",")
+      .map((url) => url.trim())
+      .join(",");
+  };
+
+  const isValidUrl = (url: string): boolean => {
+    const validURLRegex =
+      /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)/;
+    return validURLRegex.test(url);
+  };
+
+  const isInputValid = (input: string): boolean => {
+    const urlList = input.split(",");
+    return urlList.every(isValidUrl);
+  };
+
+  const calcValidUrlCount = (input: string): number => {
+    const urlList = input.split(",");
+    return urlList.filter(isValidUrl).length;
+  };
+
   return (
     <div>
       <form ref={formRef} onSubmit={submitForm}>
@@ -36,7 +59,10 @@ export const URLInput = () => {
           cols={60}
           placeholder="Enter URLs comma-separated..."
         ></textarea>
-        <TextInputStats value={userInput} />
+        <TextInputStats
+          urlCount={calcValidUrlCount(cleanInput(userInput))}
+          inputValidity={isInputValid(cleanInput(userInput))}
+        />
         <br />
         <button type="submit">Submit URLs for processing</button>
       </form>
