+++ twitch-spy-frontend/src/components/util/TextInputStats.css
@@ -1,0 +1,6 @@
+.valid {
+  color: green;
+}
+.invalid {
+  color: red;
+}
